Removed buttons from other users posts

+++ BlueBird/main/views.py
@@ -68,6 +68,10 @@ def create_room(request):
 def update_room(request, pk):
     room = Room.objects.get(id = pk)
     form = RoomForm(instance = room)
+
+    if request.user != room.host:
+        return HttpResponse("You are not authorized to view this page.")
+
     if request.method == 'POST':
         form = RoomForm(request.POST, instance = room)
         if form.is_valid():
@@ -79,6 +83,10 @@ def update_room(request, pk):
 @login_required(login_url = 'login')
 def delete_room(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse("You are not authorized to view this page.")
+
     if request.method == 'POST':
         room.delete()
         return redirect('/')
